* AutoMap facility is added.

+++ Source/Demo/Bosphorus.Dao.NHibernate.Demo.Business.Model/Bank.cs
@@ -2,7 +2,7 @@
 {
     public class Bank
     {
-        public virtual long Id { get; set; }
+        public virtual int Id { get; set; }
         public virtual string No { get; set; }
         public virtual string Name { get; set; }
     }
